\name{makeDataMethGFF}
\alias{makeDataMethGFF}

\title{GFF methylation files}
\description{
Create methData object from processed .gff files}

\usage{
makeDataMethGFF(dir,chr,start,end,meth_value)
}

\arguments{
  \item{dir}{String; The local directory where the .gff files are located}
  \item{chr}{String; Chromosome under study}
  \item{start}{Integer; The start position of genomic region under study}
  \item{end}{Integer; The end position of the genomic region under study} 
  \item{meth_value}{double; level of methylation on CpG}
}

\details{
This function reads and processes GFF files and creates a list object (like the one generated by MethAlignNW()) which can be later analyzed through the visualization, classification and clustering functions.
}

\value{
methData object
}

\author{Arie Zackay <arie.zackay@mail.huji.ac.il>, Christine Steinhoff <steinhof@molgen.mpg.de>}


\examples{
methGFF <- makeDataMethGFF(dir=system.file(package="methVisual",
			"examples/GFF"),chr="7",start=0,end=70,0.75)
methGFF
}
\keyword{graphs}


