\name{methData}
\alias{methData}

\title{BiQAnalyzer dataset}
\description{
Summary of BiQAnalyzer methylation Data created using the function MethAlignNW()
}
\usage{data(methData)}
\format{
A list
}
\details{
methData is the summary of methylation data given in the software BiqAnalyzer. the list is created by applying MethAlignNW() on BiQAnalyzer example data.
}

\references{
http://biq-analyzer.bioinf.mpi-inf.mpg.de/
}

\examples{
data(methData)
       
}
\keyword{datasets}
